requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
beautifulsoup4>=4.11.0
requests>=2.28.0
httpx>=0.24.0
python-dotenv>=1.0.0
growwapi>=0.1.0
async-lru>=2.0.0
pydantic>=2.0.0
